Add documentation and rename ArticleResult to ArticleInfo
