+++ README.md
@@ -3,129 +3,130 @@
 [![Build Status](https://api.travis-ci.com/castorini/covidex.svg?branch=master)](https://travis-ci.org/castorini/covidex)
 [![LICENSE](https://img.shields.io/badge/license-Apache-blue.svg?style=flat)](https://www.apache.org/licenses/LICENSE-2.0)
 
-This repository contains the API server, neural models, and UI client for [Covidex](https://covidex.ai), a neural search engine for the [COVID-19 Open Research Dataset (CORD-19)](https://pages.semanticscholar.org/coronavirus-research).
+This repository contains the API server, neural models, and UI client for [Covidex](https://covidex.ai), a neural search engine for the [COVID-19 Open Research Dataset (CORD-19)](https://pages.semanticscholar.org/coronavirus-research). For a description of our system, check out this paper: [Covidex: Neural Ranking Models and Keyword Search Infrastructure for the COVID-19 Open Research Dataset](https://www.aclweb.org/anthology/2020.sdp-1.5/).
 
-For a description of our system, check out this paper: [Covidex: Neural Ranking Models and Keyword Search Infrastructure for the COVID-19 Open Research Dataset](https://arxiv.org/pdf/2007.07846.pdf).
+We also provide neural search infrastructure for searching domain-specific scholarly literature via [Cydex](https://cydex.ai). This paper details the abstractions developed on top of Covidex to facilitate domain-specific search: [Cydex: Neural Search Infrastructure for the Scholarly Literature](https://www.aclweb.org/anthology/2020.sdp-1.19/).
 
-## Local Deployment
+## Environment Setup
 
-#### API Server
+### API Server
 
-Install CUDA 10.1
-+ For Ubuntu, follow [these instructions](https://developer.nvidia.com/cuda-10.1-download-archive-update2)
-+ For Debian run `sudo apt-get install nvidia-cuda-toolkit`
+1. Install CUDA 10.1
 
-Install [Anaconda](https://docs.anaconda.com/anaconda/install/linux/) (currently version 2020.02)
+- For Ubuntu, follow [these instructions](https://developer.nvidia.com/cuda-10.1-download-archive-update2)
+- For Debian run `sudo apt-get install nvidia-cuda-toolkit`
+
+2. Install [Anaconda](https://docs.anaconda.com/anaconda/install/linux/) (currently version 2020.02)
+
 ```
 wget https://repo.anaconda.com/archive/Anaconda3-2020.02-Linux-x86_64.sh
 bash Anaconda3-2020.02-Linux-x86_64.sh
 ```
 
-Install Java 11 and Maven
+3. Install Java 11 and Maven
+
 ```
 sudo apt-get install openjdk-11-jre openjdk-11-jdk maven
 ```
 
+4. Create an Anaconda environment for Python 3.7
 
-Build the [latest Anserini indices](https://github.com/castorini/anserini/blob/master/docs/experiments-cord19.md)
-```
-sh scripts/update-anserini-index.sh
-```
-
-Create an Anaconda environment for Python 3.7
 ```
 conda create -n covidex python=3.7
 ```
 
-Activate the Anaconda environment
+5. Activate the Anaconda environment
+
 ```
 conda activate covidex
 ```
 
-Install Python dependencies
+6. Install Python dependencies from inside `api/`
+
 ```
+cd api
 pip install -r api/requirements.txt
 ```
 
-Build the latest HNSW index for related article search
-```
-sh scripts/update-hnsw-index.sh
-```
+7. Setup index and environment variables
 
-Set up environment variables by copying over the defaults and modifying as needed
-```
-cp api/.env.sample api/.env
-```
+   - Build Anserini indices for your dataset. We provide instructions for setting up Covidex with both [CORD-19 and the ACL Anthology](docs/setup-index.md). Instructions to add support for new datasets is found under [docs/adding-datasets.md](docs/adding-datasets.md)
 
-Run the server (make sure you are in the `api/` folder)
-```
-uvicorn app.main:app --reload --port=8000
-```
+   - Set up environment variables by copying over the defaults from `api/.env.sample` into a new `api/.env` file, and **modifying as needed**. This requires setting the correct index and schema locations, CUDA devices, and enabling/disabling various services (highlighting, related search, neural ranking, etc.). Set `DEVELOPMENT=False` for production deployments.
 
-The server wil be running at [localhost:8000](http://localhost:8000) with API documentation at [/docs](http://localhost:8000/docs)
+### UI Client
 
+1. Install [Node.js 14+](https://nodejs.org/en/download/) and [Yarn](https://classic.yarnpkg.com/en/docs/install/).
 
-#### UI Client
+2. Install dependencies from inside `/client`
 
-Install  [Node.js 12+](https://nodejs.org/en/download/) and [Yarn](https://classic.yarnpkg.com/en/docs/install/).
-
-Install dependencies
 ```
 yarn install
 ```
 
-Start the server
+## Local Deployment
+
+Serve the UI from inside `/client`. The client will be running at [localhost:3000](http://localhost:3000).
+
 ```
 yarn start
 ```
 
-The client will be running at [localhost:3000](http://localhost:3000)
+Separately, run the API server from inside `/api`. The server wil be running at [localhost:8000](localhost:8000).
 
-
-## Production Deployment
-
-Redirect port 80 to specified port since only root can bind to port 80 (the below command uses port 8000):
 ```
-sudo iptables -t nat -A PREROUTING -p tcp --dport 80 -j REDIRECT --to-port 8000
+uvicorn app.main:app --reload --port=8000
 ```
 
-Build the [latest Anserini indices](https://github.com/castorini/anserini/blob/master/docs/experiments-cord19.md)
-```
-sh scripts/update-anserini-index.sh [DATE]
-```
+## Production deployment
 
-Build the latest HNSW index for related article search
-```
-sh scripts/update-hnsw-index.sh
-```
+We provide a script under [scripts/deploy-prod.sh](scripts/deploy-prod.sh) to start the API server and serve the UI build files. This assumes the environment is set up correctly and `api/.env` contains `DEVELOPMENT=False`.
 
 Start the server (deploys to port 8000 by default):
+
 ```
 sh scripts/deploy-prod.sh
 ```
 
-*Optional:* set the environment variable `$PORT`:
+_Optional:_ set the environment variable `PORT` to use a different port:
+
 ```
-PORT=8000 sh scripts/deploy-prod.sh
+PORT=8080 sh scripts/deploy-prod.sh
 ```
 
-Log files are available under `api/logs`, where new files are created daily based on UTC time. All filenames have the date appended except for the current one, which will be named `search.log` or `related.log`.
-
+Log files are available under `api/logs`. New files are created daily based on UTC time. All filenames have the date appended, except for the current one, which will be named `search.log` or `related.log`.
 
 ## Testing
 
-To run all API tests
+Run all API tests:
+
 ```
 TESTING=true pytest api
 ```
 
-
 ## How do I cite this work?
+
 ```
-@article{zhang2020covidex,
-  title={Covidex: Neural Ranking Models and Keyword Search Infrastructure for the COVID-19 Open Research Dataset},
-  author={Zhang, Edwin and Gupta, Nikhil and Tang, Raphael and Han, Xiao and Pradeep, Ronak and Lu, Kuang and Zhang, Yue and Nogueira, Rodrigo and Cho, Kyunghyun and Fang, Hui and others},
-  journal={arXiv preprint arXiv:2007.07846},
-  year={2020}
+@inproceedings{zhang2020covidex,
+  title = "Covidex: Neural Ranking Models and Keyword Search Infrastructure for the {COVID}-19 Open Research Dataset",
+  author = "Zhang, Edwin  and
+    Gupta, Nikhil  and
+    Tang, Raphael  and
+    Han, Xiao  and
+    Pradeep, Ronak  and
+    Lu, Kuang  and
+    Zhang, Yue  and
+    Nogueira, Rodrigo  and
+    Cho, Kyunghyun  and
+    Fang, Hui  and
+    Lin, Jimmy",
+  booktitle = "Proceedings of the First Workshop on Scholarly Document Processing",
+  month = nov,
+  year = "2020",
+  address = "Online",
+  publisher = "Association for Computational Linguistics",
+  url = "https://www.aclweb.org/anthology/2020.sdp-1.5",
+  doi = "10.18653/v1/2020.sdp-1.5",
+  pages = "31--41",
 }
 ```
