+++ scripts/update-covidex-hnsw.sh
